refactor(map): extract marker click and drag handlers
Extract marker interaction logic from overviewLayerMarkerSync into dedicated handlers module. Improve code organization by separating marker creation, event handling, and state management. Add support for resetting imported locations to original positions.

Scope: map
Change-Type: refactor
